fix 更新命名和格式錯誤
