+++ models/chatImg.js
@@ -1,0 +1,24 @@
+const mongoose = require('mongoose')
+const chatImgSchema = new mongoose.Schema(
+  {
+    name: {
+      type: String
+    },
+    img: {
+      type: String,
+      default: []
+    },
+    createdAt: {
+      type: Date,
+      default: Date.now
+      // select: false
+    }
+  },
+  {
+    versionKey: false,
+    new: true
+  }
+)
+const chatImg = mongoose.model('chatImg', chatImgSchema)
+
+module.exports = chatImg
